Make COSE deserialization more robust
Previously, COSE keys could only be deserialized if the map entries were
in the correct order. With this patch, the deserialization is made more
robust (and also more concise) by introducing a helper struct,
RawPublicKey, that is used as an intermediate step during the
deserialization. Also, we make the algorithm field optional to conform
with the spec.

This is a port from ctap-types:
    https://github.com/solokeys/ctap-types/pull/8